+++ src/styles/App.css
@@ -1,0 +1,12 @@
+.App {
+  text-align: center;
+  font-family: "Poppins";
+}
+
+button {
+  font-family: "Poppins";
+}
+
+.logo {
+  height: 30px;
+}
