+++ src/styles/Filter.css
@@ -1,0 +1,6 @@
+.button-list {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+  }
